requirements.txt
solara
sparkfun-qwiic-relay
adafruit-blinka
lgpio
circuitpython-stts22h
aioesphomeapi
paho-mqtt
aiomqtt

